Shares auction

+++ shared/src/commonMain/kotlin/ua/vald_zx/game/rat/race/card/shared/Cards.kt
@@ -104,7 +104,7 @@ sealed class BoardCard(val type: BoardCardType) {
         data class Shares(
             val description: String,
             val price: Long,
-            val maxCount: Int,
+            val maxCount: Long,
             val sharesType: SharesType
         ) : Chance()
